test(windows): tighten windowsHide spawn assertions and match test naming convention
